git_summarizer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.5
langchain-core==0.1.16
langchain-google-genai==0.0.5
gitpython==3.1.40
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
jinja2==3.1.2
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
attrs==23.1.0
cffi==1.15.1
cryptography==41.0.7
platformdirs==2.6.2
pytz==2023.3
google-cloud-aiplatform==1.42.1